Console mode settings in settings.conf + don't save in settings.conf if no file system.

+++ Aura Operating System/Aura OS/System/Utils/Settings.cs
@@ -34,7 +34,10 @@ namespace Aura_OS.System.Utils
         /// </summary>
         public static void PushValues()
         {
-            File.WriteAllLines(@"0:\System\settings.conf", file);
+            if (Kernel.SystemExists)
+            {
+                File.WriteAllLines(@"0:\System\settings.conf", file);
+            }
         }
 
         /// <summary>
